refactor: Move cpp interface in its own crate
