--- Justfile
@@ -1,3 +1,0 @@
-cexample:
-  cmake -B build/
-  make -C build/; ./build/cpp_with_rust
