+++ qibo-core-cxx/.gitignore
@@ -1,0 +1,2 @@
+# CMake
+build/
