+++ qibo-core-cxx/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "qibo-core-cxx"
+
+authors.workspace = true
+description.workspace = true
+readme.workspace = true
+categories.workspace = true
+edition.workspace = true
+keywords.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[lib]
+name = "qibo_core"
+crate-type = ["staticlib"]
+
+[dependencies]
+cxx = "1.0"
+qibo-core = { version = "0.0.1", path = ".." }
+
+[build-dependencies]
+cxx-build = "1.0"
