+++ qibo-core-cxx/Justfile
@@ -1,0 +1,4 @@
+cexample:
+  cmake -B build/
+  make -C build/
+  ./build/circuit
